+++ src/main/java/net/preibisch/stitcher/leap/MotionHandler.java
@@ -3,9 +3,9 @@ package net.preibisch.stitcher.leap;
 public interface MotionHandler {
     void scale(double scale);
 
-    void translate(float[] vector);
+    void translate(double[] vector);
 
-    void rotate(float[] vector);
+    void rotate(double[] vector);
 
     void reset();
 }
